SSH ControlMaster
Update how the Connection object works by utilizing ControlMaster option
and change the way execution works to always subshell to a new ssh
proceess that shares the socket to execute the command. This greatly
simplifies (and potentially introduces a deadlock situation when
commands emit significant output, but we'll deal with that later)
everything around command execution and (hopefully) introduces an API
that might scale beyond subshelling to the SSH command for remote
execution.